Move all components in a directory
