+++ packages/twenty-front/src/modules/activities/emails/right-drawer/components/RightDrawerThread.tsx
